requirements.txt
asgiref==3.5.2
astroid==2.11.7
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.5
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
filelock==3.7.1
identify==2.5.2
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
oauthlib==3.2.0
packaging==21.3
platformdirs==2.5.2
pre-commit==2.20.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
uritemplate==4.1.1
urllib3==1.26.10
virtualenv==20.16.2
wrapt==1.14.1
